Check userdefaults in local/global screen
